# filafood-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Filafood API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/commands": {
            "get": {
                "tags": [
                    "Commands"
                ],
                "summary": "Obter todos os recursos",
                "operationId": "2bd312f2ce5c8bfe3c98bc111295705c",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Quantidade de dados por página",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "identify",
                        "in": "query",
                        "description": "Filtrar recurso pelo número identificador",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Filtrar recurso pelo número da comanda",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Armazenar novo recurso no banco de dados",
                "operationId": "ecba06f6e3ea54e9d22532bf6639b008",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number": {
                                        "type": "integer"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "ID do produto",
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                                                },
                                                "quantity": {
                                                    "description": "Quantidade do produto",
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/commands/{id}": {
            "get": {
                "tags": [
                    "Commands"
                ],
                "summary": "Ver um recurso",
                "operationId": "75cbb479ab9550224982fc9880c56287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obter todos os recursos",
                "operationId": "b63ec4394c475fe1a0fdc836e1e1ed99",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Quantidade de dados por página",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Filtrar recurso pelo número da comanda",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Ver um recurso",
                "operationId": "c82bf974ea99be4d0ad4491f2e9a3096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}